+++ CHANGELOG.md
@@ -6,6 +6,9 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+
+### Added
+- new field addressType in checkout
 ## [0.1.8] - 2022-07-16
 ### Changed
 - rename CTA `Continue to shipping` to `Save address`